api/requirements.txt
Flask==2.1.0
Flask_Cors==4.0.0
gunicorn==20.1.0
google-cloud-datastore==2.16.1
